+++ NEWS.md
@@ -1,6 +1,7 @@
 ## version 0.2.17
 ### Features
 - adds informed prior distributions for dichotomous and time to event outcomes based on Cochrane Database of Systematic Reviews to `prior_informed()` function
+- adds bridge object conveneience function `bridge_object()` (fixes: https://github.com/FBartos/BayesTools/issues/28)
 
 ### Fixes
 - ability to run more than 4 chains (fixes: https://github.com/FBartos/BayesTools/issues/20)